requirements.txt
django==1.8.4
feincms==1.11.1
django-model-utils==2.3.1


https://github.com/AASHE/aashe-auth/archive/master.zip
